requirements.txt
flask
pandas
numpy
joblib
scikit-learn
openai
python-dotenv
gunicorn
requests
pydantic